requirements.txt
amqp==5.1.1
appnope==0.1.3
APScheduler==3.6.3
arrow==1.2.2
asgiref==3.5.1
asn1crypto==1.5.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
billiard==3.6.4.0
bip-utils==2.3.0
bip32==3.1
bip32utils==0.3.post4
bit==0.8.0
cachetools==4.2.2
captcha==0.4
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coincurve==17.0.0
crcmod==1.7
decorator==5.1.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.13
django-celery-beat==2.2.1
django-cors-headers==3.12.0
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-timezone-field==4.2.3
docker-py==1.8.0
ecdsa==0.17.0
ed25519-blake2b==1.4
executing==0.8.3
flake8==4.0.1
funcsigs==1.0.2
future==0.18.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
ipython==8.3.0
jedi==0.18.1
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==4.5.1
kombu==5.2.4
matplotlib-inline==0.1.3
mccabe==0.6.1
mnemonic==0.13
more-itertools==8.13.0
multicolorcaptcha==1.2.0
packaging==21.3
parso==0.8.3
path==16.4.0
path.py==12.5.0
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.0.0a3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-sr25519-bindings==0.1.4
pyaes==1.6.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pyflakes==2.4.0
Pygments==2.12.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyTelegramBotAPI==4.5.1
pyTelegramBotCAPTCHA==1.1.4
pytest==7.1.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.3.1
requests==2.27.1
scalecodec==1.0.35
sha256==0.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tabulate==0.8.9
tomli==2.0.1
tornado==6.1
traitlets==5.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.2
whitenoise==6.1.0
wrapt==1.14.1
